backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
requests
beautifulsoup4
python-dotenv
httpx
